requirements.txt
Django==2.2
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
PyJWT==2.1.0
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
uritemplate==3.0.1
urllib3==1.26.5
